# Reuse heap elements for Push/Pop operations