Optimize memory usage by reusing heap elements for Push/Pop operations described [here](https://lemire.me/blog/2019/08/16/faster-threshold-queries-with-cache-sensitive-scancount/)

```
benchmark                         old ns/op     new ns/op     delta
BenchmarkSuggest-8                21650         21742         +0.42%
BenchmarkSuggestWordsOnDisc-8     332217        305817        -7.95%

benchmark                         old allocs     new allocs     delta
BenchmarkSuggestWordsOnDisc-8     636            611            -3.93%

benchmark                         old bytes     new bytes     delta
BenchmarkSuggestWordsOnDisc-8     25371         22465         -11.45%
```